Introduce a common mistake with an example of Indirection Without Abstraction
