+++ src/main/java/com/ite/itea/persistence/FileSystemUserRepository.java
@@ -1,48 +1,11 @@
 package com.ite.itea.persistence;
 
 import com.ite.itea.domain.user.User;
-import com.ite.itea.domain.user.UserRepository;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
-public class FileSystemUserRepository implements UserRepository {
+public interface FileSystemUserRepository {
 
-    private final File usersFile;
-
-    public FileSystemUserRepository(File usersFile) {
-        this.usersFile = usersFile;
-    }
-
-    public List<User> all() {
-        return parseUsersFromFile(usersFile);
-    }
-
-    public Optional<User> byLastName(String lastName) {
-        return parseUsersFromFile(usersFile).stream()
-                .filter(user -> user.lastname().equals(lastName))
-                .findFirst();
-    }
-
-    private List<User> parseUsersFromFile(File usersFile) {
-        try {
-            final var lines = Files.readAllLines(usersFile.toPath());
-            return lines.stream()
-                    .map(this::parseUser)
-                    .collect(Collectors.toList());
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private User parseUser(String line) {
-        String[] userFromFile = line.split(" ");
-        String[] ordersFromFile = userFromFile[2].split(",");
-        return new User(userFromFile[0], userFromFile[1], Arrays.asList(ordersFromFile));
-    }
+    List<User> all();
+    Optional<User> byLastName(String lastName);
 }
